Fix the error message when loading the default `skaffold.yaml`
