+++ pkg/skaffold/util/util.go
@@ -150,30 +150,6 @@ func StringPtr(s string) *string {
 	return &o
 }
 
-func ReadConfiguration(filename string) ([]byte, error) {
-	switch {
-	case filename == "":
-		return nil, errors.New("filename not specified")
-	case filename == "-":
-		return ioutil.ReadAll(os.Stdin)
-	case IsURL(filename):
-		return Download(filename)
-	default:
-		directory := filepath.Dir(filename)
-		baseName := filepath.Base(filename)
-		if baseName != "skaffold.yaml" {
-			return ioutil.ReadFile(filename)
-		}
-		contents, err := ioutil.ReadFile(filename)
-		if err != nil {
-			logrus.Infof("Could not open skaffold.yaml: \"%s\"", err)
-			logrus.Infof("Trying to read from skaffold.yml instead")
-			return ioutil.ReadFile(filepath.Join(directory, "skaffold.yml"))
-		}
-		return contents, err
-	}
-}
-
 func IsURL(s string) bool {
 	return strings.HasPrefix(s, "http://") || strings.HasPrefix(s, "https://")
 }
